Replace FirstOrDefault with Find in various methods
In `ArticleDefinitionFactory.cs`, `FeedController.cs`, and `GetFeedHandler.cs`, the `FirstOrDefault` method used to retrieve the first matching item based on a key has been replaced with the `Find` method. This change was made to throw an exception if no matching item is found, as `Find` throws an exception in such cases, while `FirstOrDefault` simply returns null.

http://sonarqube:9000/coding_rules?open=csharpsquid%3AS6602&rule_key=csharpsquid%3AS6602